+++ tests/autocontext_index.test.js
@@ -2,7 +2,7 @@ process.env.NO_GIT = "true";
 const fs = require('fs');
 const path = require('path');
 const assert = require('assert');
-const { load_context_from_index, setMemoryRepo } = require('../memory');
+const { load_context_from_index, setMemoryRepo } = require('../src/memory');
 const { contextFilename } = require('../logic/memory_operations');
 
 async function run() {
